daniele_marco/daniele/_old_root_files/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
pydantic==2.5.3
sqlalchemy==2.0.25
asyncpg==0.29.0
redis==5.0.1
chromadb==0.4.22
sentence-transformers==2.3.1
dspy-ai==2.1.0
pdfplumber==0.10.3
python-docx==1.1.0
aiofiles==23.2.1
python-multipart==0.0.6
